+++ packages/sdk/src/utils/paymentTokens.ts
@@ -615,4 +615,21 @@ export const chainPaymentTokensMap = {
       decimals: 18,
     },
   ],
+  // Redstone Testnet
+  17001: [
+    {
+      chainId: 17001,
+      address: zeroAddress,
+      symbol: 'ETH',
+      name: 'ETH',
+      decimals: 18,
+    },
+    {
+      chainId: 17001,
+      address: '0x09d940117bd1df3165a22820517a946426605f60',
+      symbol: 'WETH',
+      name: 'WETH',
+      decimals: 18,
+    },
+  ],
 } as Record<number, PaymentToken[]>