+++ packages/ethers-wallet-adapter/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @reservoir0x/ethers-wallet-adapter
 
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies [9c46f40]
+  - @reservoir0x/reservoir-sdk@2.0.11
+
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies [2198ecd]
+  - @reservoir0x/reservoir-sdk@2.0.10
+
 ## 1.0.9
 
 ### Patch Changes