+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,17 @@
 ## 1.12.15
 
+## 2.0.11
+
+### Patch Changes
+
+- 9c46f40: Add redstone testnet configuration
+
+## 2.0.10
+
+### Patch Changes
+
+- 2198ecd: Add apex pop testnet configuration
+
 ## 2.0.9
 
 ### Patch Changes