+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,21 @@
 ## 1.25.10
 
+## 2.4.3
+
+### Patch Changes
+
+- 9c46f40: Add redstone testnet configuration
+- Updated dependencies [9c46f40]
+  - @reservoir0x/reservoir-sdk@2.0.11
+
+## 2.4.2
+
+### Patch Changes
+
+- 2198ecd: Add apex pop testnet configuration
+- Updated dependencies [2198ecd]
+  - @reservoir0x/reservoir-sdk@2.0.10
+
 ## 2.4.1
 
 ### Patch Changes